refactored user model to use service file
